Move existing docs into doc.rs
